requirements.txt
coreapi==2.0.8
Django==1.10.2
django-cors-headers==1.1.0
django-rest-swagger==2.1.0
djangorestframework==3.5.1
djangorestframework-jwt==1.8.0
django-nose
python-coveralls
mysqlclient==1.3.7
openapi-codec==1.1.7
uritemplate==3.0.0
uritemplate.py==0.3.0
urllib3==1.13.1
uWSGI==2.0.13.1
virtualenv==15.0.2
mkdocs
drf-nested-routers
django-filter
drf-haystack
elasticsearch
